requirements.txt
asgiref==3.7.2
Django==4.2.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-standardized-errors==0.12.5
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
